core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "TaxiWatch Backend - Real-Time Taxi Monitoring System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "daphne>=4.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "boto3>=1.28.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = []
